czso: Use Open Data from the Czech Statistical Office in R

Get programmatic access to the open data provided by the Czech Statistical Office (CZSO, https://czso.cz).

See also

Author

Maintainer: Petr Bouchal pbouchal@gmail.com (ORCID)

Other contributors:

Jindra Lacko [contributor]