Filter the catalogue using a set of keywords

Filter the catalogue using a set of keywords

czso_filter_catalogue(catalogue, search_terms)

Arguments

catalogue

a catalogue as returned by czso_get_catalogue()

search_terms

#' A regex pattern (incl. plain text), or a vector of regex patterns, to filter the catalogue by. A case-insensitive filter is performed on the title, description and keywords. The search returns only catalogue entries where all the patterns are matched anywhere within the title, description or keywords.

Value

A tibble with the filtered catalogue.

See also

Examples

ctlg <- czso_get_catalogue()
czso_filter_catalogue(ctlg, search_terms = c("kraj", "me?zd"))
#> # A tibble: 1 × 12
#>   dataset_iri     dataset_id title provider description spatial modified   page 
#>   <chr>           <chr>      <chr> <chr>    <chr>       <chr>   <date>     <chr>
#> 1 https://vdb.cz… 110080     Prům… Český s… Datová sad… https:… 2024-06-28 http…
#> # ℹ 4 more variables: periodicity <chr>, start <date>, end <date>,
#> #   keywords_all <chr>
czso_filter_catalogue(ctlg, search_terms = c("úmrt", "orp"))
#> # A tibble: 3 × 12
#>   dataset_iri     dataset_id title provider description spatial modified   page 
#>   <chr>           <chr>      <chr> <chr>    <chr>       <chr>   <date>     <chr>
#> 1 https://vdb.cz… 130140r15  Nadě… Český s… Datová sad… https:… 2024-06-28 http…
#> 2 https://vdb.cz… 130140r16  Nadě… Český s… Datová sad… https:… 2024-06-28 http…
#> 3 https://vdb.cz… 130140r17  Nadě… Český s… Datová sad… https:… 2024-06-28 http…
#> # ℹ 4 more variables: periodicity <chr>, start <date>, end <date>,
#> #   keywords_all <chr>
czso_filter_catalogue(ctlg, search_terms = c("kraj", "vazba", "orp"))
#> # A tibble: 2 × 12
#>   dataset_iri     dataset_id title provider description spatial modified   page 
#>   <chr>           <chr>      <chr> <chr>    <chr>       <chr>   <date>     <chr>
#> 1 https://vdb.cz… cis100vaz… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#> 2 https://vdb.cz… cis108vaz… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#> # ℹ 4 more variables: periodicity <chr>, start <date>, end <date>,
#> #   keywords_all <chr>
czso_filter_catalogue(ctlg, search_terms = c("ISCO", "číselník"))
#> # A tibble: 9 × 12
#>   dataset_iri     dataset_id title provider description spatial modified   page 
#>   <chr>           <chr>      <chr> <chr>    <chr>       <chr>   <date>     <chr>
#> 1 https://vdb.cz… cis5705va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#> 2 https://vdb.cz… cis5706va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#> 3 https://vdb.cz… cis5707va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#> 4 https://vdb.cz… cis5708va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#> 5 https://vdb.cz… cis5705    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 6 https://vdb.cz… cis5706    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 7 https://vdb.cz… cis5707    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 8 https://vdb.cz… cis5708    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 9 https://vdb.cz… cis5709    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> # ℹ 4 more variables: periodicity <chr>, start <date>, end <date>,
#> #   keywords_all <chr>
czso_filter_catalogue(ctlg, search_terms = c("zaměstnání", "číselník"))
#> # A tibble: 16 × 12
#>    dataset_iri    dataset_id title provider description spatial modified   page 
#>    <chr>          <chr>      <chr> <chr>    <chr>       <chr>   <date>     <chr>
#>  1 https://vdb.c… cis5661va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  2 https://vdb.c… cis5662va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  3 https://vdb.c… cis5663va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  4 https://vdb.c… cis5705va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  5 https://vdb.c… cis5706va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  6 https://vdb.c… cis5707va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  7 https://vdb.c… cis5708va… Vazb… Český s… Vazba mezi… https:… 2023-03-03 http…
#>  8 https://vdb.c… cis5661    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#>  9 https://vdb.c… cis5662    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 10 https://vdb.c… cis5663    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 11 https://vdb.c… cis5664    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 12 https://vdb.c… cis5705    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 13 https://vdb.c… cis5706    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 14 https://vdb.c… cis5707    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 15 https://vdb.c… cis5708    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> 16 https://vdb.c… cis5709    Číse… Český s… Klasifikac… https:… 2023-03-03 http…
#> # ℹ 4 more variables: periodicity <chr>, start <date>, end <date>,
#> #   keywords_all <chr>