Get documentation for CZSO dataset

Retrieves the URL/downloads the file containing the documentation of the dataset, in the required format.

Usage

czso_get_dataset_doc(
  dataset_id,
  action = c("return", "open", "download"),
  destfile = NULL,
  format = c("html", "pdf", "word")
)

Arguments

dataset_id

Dataset ID

action

Whether to return URL (the default), download the file, or open the URL in the default web browser.

destfile

Where to save the file. Only used if if action = download.

format

What file format to access: html (the default), pdf, or word.

Value

if action = download, the path to the downloaded file; file URL otherwise.

Details

The document to which this functions provides access contains methodological background on the specified dataset and is identified by the schema field in the list returned by czso_get_dataset_metadata().

See also

Examples

# \donttest{
czso_get_dataset_doc("110080")
#> [1] "https://csu.gov.cz/docs/107508/f3dcc902-8fc9-0b87-7aa2-2e560f5830df/110080-25schema052325.json?version=1.0"
# }