Get catalogue of open CZSO datasets

Downloads and processes a list of all registered Czech open data datasets, returning (by default) those accessible through get_table() from the CZSO.

get_czso_catalogue(
  provider = "\\u010cesk\\u00fd statistick\\u00fd \\u00fa\\u0159ad",
  title_filter = NULL,
  description_filter = NULL,
  keyword_filter = NULL,
  provider_filter = NULL,
  force_redownload = F
)

Arguments

provider
character, can be of length > 1. Provider to select for. Defaults to (the Czech name of) CZSO. Must be exact match. If set to NULL, returns full list of all datasets.

title_filter
character, text to use for filtering the set by title. Case sensitive. Can be a regular expression.

description_filter
character, text to use for filtering the set by description. Case sensitive. Can be a regular expression.

keyword_filter
character, text to use for filtering the set by keyword. Case sensitive. Can be a regular expression.

provider_filter
character, text to use for filtering the set by provider Case sensitive. Can be a regular expression.

force_redownload
integer. Whether to redownload data source file even if already cached. Defaults to FALSE.

Value

a data frame. If provider param is left to default, contains a column called czso_id, which can be used as dataset_id parameter in get_table().

Details

If provider is NULL, returns the whole list, without CZSO-specific identifier usable in get_table().

If provider is left unset, returns data frame listing CZSO's datasets, with a czso_id column usable in get_table.

Other values of provider must be exact matches. Use provider_filter for text/regex matching.

All *_filter arguments are case sensitive and can be regular expressions.

Original data: https://data.gov.cz/soubor/datov%C3%A9-sady.csv

See also

Examples

if (FALSE) {
get_czso_catalogue()
get_czso_catalogue(NULL)
get_czso_catalogue(title_filter = "[Mm]zd[ay]")
get_czso_catalogue(provider = "Ministerstvo vnitra")
get_czso_catalogue(provider_filter = "[Mm]inisterstvo")
}