Get catalogue of open CZSO datasets

Retrieves a list of all CZSO's open datasets available from the Czech Open data catalogue.

czso_get_catalogue()

Value

a data frame with details on all CZSO datasets available in the Czech National Open Data Catalogue. The columns are fairly well described by their names, except:

some columns contain IRIs instead of human readable text; still you can deduce the content from the IRI.

the spatial columns contains an IRI ending in the pattern unit_type/unit_code. The unit_type denotes what unit the data covers (scope/domain not granularity) and the second identifies the unit covered. The unit_type will usually be "stat" for "state" and the unit_code will be 1. The unit_type can also be "KR" for region or "OB" for municipality, or "OK" for district. In that case, the unit_code will be a code of that unit.

page points to the documentation, i.e. methodology notes for the dataset.

Details

Use the dataset_id column as an argument to get_czso_table().

See also

Examples

# \donttest{
czso_get_catalogue()

#> ℹ Reading data from data.gov.cz

#> ✔ Done downloading and reading data

#> ℹ Transforming data

#> # A tibble: 613 x 14
#>    dataset_iri dataset_id title provider description spatial temporal
#>    <chr>       <chr>      <chr> <chr>    <chr>       <chr>   <chr>   
#>  1 https://da… 060003     Vyba… Český s… Datová sad… https:… https:/…
#>  2 https://da… 270229     Skli… Český s… Datová sad… https:… https:/…
#>  3 https://da… 110080     Prům… Český s… Datová sad… https:… https:/…
#>  4 https://da… 270230     Hosp… Český s… Datová sad… https:… https:/…
#>  5 https://da… 340130     Veli… Český s… Datová sad… https:… https:/…
#>  6 https://da… cis203     Číse… Český s… Druh lesní… https:… https:/…
#>  7 https://da… cis199     Číse… Český s… Druh činno… https:… https:/…
#>  8 https://da… cis202     Číse… Český s… Druh lesní… https:… https:/…
#>  9 https://da… cis7605    Číse… Český s… Specifikac… https:… https:/…
#> 10 https://da… cis206     Číse… Český s… Příčina na… https:… https:/…
#> # … with 603 more rows, and 7 more variables: modified <dttm>, page <chr>,
#> #   periodicity <chr>, periodicity_abb <chr>, start <date>, end <date>,
#> #   keywords <chr>

# }