Get catalogue of open CZSO datasets

Retrieves a list of all CZSO's open datasets available from the Czech Open data catalogue.

czso_get_catalogue()

Value

a data frame with details on all CZSO datasets available in the Czech National Open Data Catalogue. The columns are fairly well described by their names, except:

some columns contain IRIs instead of human readable text; still you can deduce the content from the IRI.

the spatial columns contains an IRI ending in the pattern unit_type/unit_code. The unit_type denotes what unit the data covers (scope/domain not granularity) and the second identifies the unit covered. The unit_type will usually be "stat" for "state" and the unit_code will be 1. The unit_type can also be "KR" for region or "OB" for municipality, or "OK" for district. In that case, the unit_code will be a code of that unit.

page points to the documentation, i.e. methodology notes for the dataset.

Details

Pass the string in the dataset_id column to get_czso_table(). dataset_iri is the unique identifier of the dataset in the national catalague and also the URL containing all metadata for the dataset.

See also

Examples

# \donttest{
czso_get_catalogue()
#> # A tibble: 724 × 12
#>    dataset_iri    dataset_id title provider description spatial modified   page 
#>    <chr>          <chr>      <chr> <chr>    <chr>       <chr>   <date>     <chr>
#>  1 https://vdb.c… 170242     Bila… Český s… Datová sad… https:… 2021-07-27 http…
#>  2 https://vdb.c… 290038r05  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  3 https://vdb.c… 290038r06  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  4 https://vdb.c… 290038r07  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  5 https://vdb.c… 290038r08  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  6 https://vdb.c… 290038r09  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  7 https://vdb.c… 290038r10  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  8 https://vdb.c… 290038r11  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#>  9 https://vdb.c… 290038r12  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#> 10 https://vdb.c… 290038r13  Cizi… Český s… Datová sad… https:… 2021-07-28 http…
#> # … with 714 more rows, and 4 more variables: periodicity <chr>, start <date>,
#> #   end <date>, keywords_all <chr>
# }