Get dataset metadata

Get metadata from CZSO API, which can be somewhat more detailed/readable than what is provided in the dataset's entry in the output of czso_get_dataset().

Usage

czso_get_dataset_metadata(dataset_id)

Arguments

dataset_id

Dataset ID

Value

a list with elements named in English, where the names are mostly self-explanatory. So are the contents where these are dates; title, description, notes and tags only exist in Czech as far as I know. Some fields merit explanation:

resources: a list of files available to download in this dataset

frequency: see https://project-open-data.cio.gov/iso8601_guidance/ for a key

ruian_type: what type of spatial unit the data covers (spatial domain/extent/scope, not granularity). ST means "state" (this is almost always the case), "KR" means region (kraj), "OK" district (okres), "OB" municipality (obec); "RS" cohesion region (region soudržnosti, larger than region)

ruian_code: the code of the unit the data covers as per the RUIAN taxonomy

schema points to documentation while describedBy points to the technical schema in JSON or XML.

Details

As far as I can tell there is no way to get the metadata in English, though some key datasets, such as codelists, do have English-language documentation. See czso_get_table() for how to access English-language codelists (registers).

See also

Examples

# \donttest{
czso_get_dataset_metadata("110080")
#> $`@context`
#> [1] "https://ofn.gov.cz/dcat-ap-cz-rozhraní-katalogů-otevřených-dat/2024-05-28/kontexty/rozhraní-katalogů-otevřených-dat.jsonld"
#> 
#> $iri
#> [1] "https://vdb.czso.cz/pll/eweb/lkod_ld.datova_sada?id=110080"
#> 
#> $typ
#> [1] "Datová sada"
#> 
#> $název
#> $název$cs
#> [1] "Průměrná hrubá měsíční mzda a medián mezd v krajích"
#> 
#> 
#> $popis
#> $popis$cs
#> [1] "Datová sada obsahuje časovou řadu průměrných měsíčních mezd a mediánu mezd v krajích České republiky od roku 2011 včetně údajů podle pohlaví."
#> 
#> 
#> $prvek_rúian
#> [1] "https://linked.cuzk.cz/resource/ruian/stat/1"
#> 
#> $časové_pokrytí
#> $časové_pokrytí$typ
#> [1] "Časový interval"
#> 
#> $časové_pokrytí$začátek
#> [1] "2011-01-01"
#> 
#> $časové_pokrytí$konec
#> [1] "2024-12-31"
#> 
#> 
#> $klíčové_slovo
#> $klíčové_slovo$cs
#> [1] "Mzda"                "průměrná mzda"       "medián mezd"        
#> [4] "hrubá měsíční mzda."
#> 
#> 
#> $periodicita_aktualizace
#> [1] "http://publications.europa.eu/resource/authority/frequency/ANNUAL"
#> 
#> $dokumentace
#> [1] "https://csu.gov.cz/docs/107508/159084c6-c6c7-d955-2ef9-fd89e33d7800/110080-22dds.htm"
#> 
#> $téma
#> [1] "http://publications.europa.eu/resource/authority/data-theme/ECON"
#> [2] "http://publications.europa.eu/resource/authority/data-theme/SOCI"
#> [3] "http://publications.europa.eu/resource/authority/data-theme/GOVE"
#> 
#> $specifikace
#> [1] "https://opendata.gov.cz/standardy:technicke-standardy-pro-datove-sady-na-stupni-3"
#> 
#> $časové_rozlišení
#> [1] "R/P1Y"
#> 
#> $kontaktní_bod
#> $kontaktní_bod$typ
#> [1] "Organizace"
#> 
#> $kontaktní_bod$jméno
#> $kontaktní_bod$jméno$cs
#> [1] "Martin Černý, ČSÚ"
#> 
#> 
#> $kontaktní_bod$`e-mail`
#> [1] "mailto:martin.cerny@czso.cz"
#> 
#> 
#> $poskytovatel
#> [1] "https://rpp-opendata.egon.gov.cz/odrpp/zdroj/orgán-veřejné-moci/00025593"
#> 
#> $distribuce
#>          typ
#> 1 Distribuce
#>                                                                       iri
#> 1 https://vdb.czso.cz/pll/eweb/lkod_ld.datova_sada?id=110080/distribuce/0
#>                  cs         podmínky_užití.typ
#> 1 Data do roku 2024 Specifikace podmínek užití
#>                                   podmínky_užití.autorské_dílo
#> 1 https://data.gov.cz/podmínky-užití/neobsahuje-autorská-díla/
#>                                   podmínky_užití.databáze_jako_autorské_dílo
#> 1 https://data.gov.cz/podmínky-užití/není-autorskoprávně-chráněnou-databází/
#>                                          podmínky_užití.databáze_chráněná_zvláštními_právy
#> 1 https://data.gov.cz/podmínky-užití/není-chráněna-zvláštním-právem-pořizovatele-databáze/
#>                                   podmínky_užití.osobní_údaje
#> 1 https://data.gov.cz/podmínky-užití/neobsahuje-osobní-údaje/
#>                                                                                         soubor_ke_stažení
#> 1 https://csu.gov.cz/docs/107508/9643ae62-f3a1-75c4-1ebd-23c9cdb5d81a/110080-25data052325.csv?version=1.0
#>                                                                                            přístupové_url
#> 1 https://csu.gov.cz/docs/107508/9643ae62-f3a1-75c4-1ebd-23c9cdb5d81a/110080-25data052325.csv?version=1.0
#>                                              typ_média
#> 1 http://www.iana.org/assignments/media-types/text/csv
#>                                                           formát
#> 1 http://publications.europa.eu/resource/authority/file-type/CSV
#>                                                                                                       schéma
#> 1 https://csu.gov.cz/docs/107508/f3dcc902-8fc9-0b87-7aa2-2e560f5830df/110080-25schema052325.json?version=1.0
#> 
# }