Get dataset metadata

Get metadata from CZSO API, which can be somewhat more detailed/readable than what is provided in the dataset's entry in the output of czso_get_dataset().

czso_get_dataset_metadata(dataset_id)

Arguments

dataset_id
Dataset ID

Value

a list with elements named in English, where the names are mostly self-explanatory. So are the contents where these are dates; title, description, notes and tags only exist in Czech as far as I know. Some fields merit explanation:

resources: a list of files available to download in this dataset

frequency: see https://project-open-data.cio.gov/iso8601_guidance/ for a key

ruian_type: what type of spatial unit the data covers (spatial domain/extent/scope, not granularity). ST means "state" (this is almost always the case), "KR" means region (kraj), "OK" district (okres), "OB" municipality (obec); "RS" cohesion region (region soudržnosti, larger than region)

ruian_code: the code of the unit the data covers as per the RUIAN taxonomy

schema points to documentation while describedBy points to the technical schema in JSON or XML.

Details

As far as I can tell there is no way to get the metadata in English, though some key datasets, such as codelists, do have English-languge documentation. See czso_get_table() for how to access English-language codelists (registers).

See also

Examples

# \donttest{
czso_get_dataset_metadata("110080")

#> $name
#> [1] "110080"
#> 
#> $title
#> [1] "Průměrná hrubá měsíční mzda a medián mezd v krajích"
#> 
#> $notes
#> [1] "Datová sada obsahuje časovou řadu průměrných měsíčních mezd a mediánu mezd v krajích České republiky od roku 2011 včetně údajů podle pohlaví."
#> 
#> $frequency
#> [1] "R/P1Y"
#> 
#> $ruian_type
#> [1] "ST"
#> 
#> $ruian_code
#> [1] "1"
#> 
#> $publisher_uri
#> [1] "https://www.czso.cz"
#> 
#> $publisher_name
#> [1] "Český statistický úřad"
#> 
#> $maintainer_email
#> [1] "vdb@czso.cz"
#> 
#> $maintainer
#> [1] "Český statistický úřad"
#> 
#> $schema
#> [1] "https://www.czso.cz/documents/62353418/109720808/110080-19dds.htm"
#> 
#> $temporal_start
#> [1] "2011-01-01"
#> 
#> $temporal_end
#> [1] "2019-12-31"
#> 
#> $tags
#>          display_name                name
#> 1                Mzda                Mzda
#> 2       průměrná mzda       průměrná mzda
#> 3         medián mezd         medián mezd
#> 4 hrubá měsíční mzda. hrubá měsíční mzda.
#> 
#> $resources
#>                                                                        url
#> 1 https://www.czso.cz/documents/62353418/123503145/110080-20data060120.csv
#>     format                                                 license_link
#> 1 text/csv https://portal.gov.cz/portal/ostatni/volny-pristup-k-ds.html
#>                name
#> 1 Data do roku 2019
#>                                                                                                                                     description
#> 1 Datová sada obsahuje časovou řadu průměrných měsíčních mezd a mediánu mezd v krajích České republiky od roku 2011 včetně údajů podle pohlaví.
#>                                                                   describedBy
#> 1 https://www.czso.cz/documents/62353418/123503145/110080-20schema060120.json
#>    describedByType temporal_start temporal_end
#> 1 application/json     2011-01-01   2019-12-31
#> 

# }