Get CZSO table schema

Retrieves and parses the schema for the table identified by dataset_id and resource_num.

czso_get_table_schema(dataset_id, resource_num = 1)

Arguments

dataset_id
Dataset ID

resource_num
Resource number, typically 1 in CZSO (the default)

Value

a tibble with a description of the table columns, with the following items:

name: the column name.

titles: usually the duplicate of name

dc:description: a Czech-language description of the column

required: whether the column is required

datatatype: the data type of the column; either "number" or "string"

Details

Currently only handles JSON schema files for CSV files. If the schema is a different format, an error is returned pointing the user to the URL of the file.

See also

Examples

# \donttest{
czso_get_table_schema("110080")

#> # A tibble: 14 x 5
#>    name       titles     `dc:description`                      required datatype
#>    <chr>      <chr>      <chr>                                 <lgl>    <chr>   
#>  1 idhod      idhod      "unikátní identifikátor údaje Veřejn… TRUE     string  
#>  2 hodnota    hodnota    "zjištěná hodnota"                    TRUE     number  
#>  3 stapro_kod stapro_kod "kód statistické proměnné ze systému… TRUE     string  
#>  4 spkvantil… spkvantil… "kód číselníku pro kvantil"           TRUE     string  
#>  5 spkvantil… spkvantil… "kód položky z číselníku pro kvantil" TRUE     string  
#>  6 pohlavi_c… pohlavi_c… "kód číselníku pro pohlaví"           TRUE     string  
#>  7 pohlavi_k… pohlavi_k… "kód položky číselníku pro pohlaví"   TRUE     string  
#>  8 rok        rok        "rok referenčního období ve formátu … TRUE     number  
#>  9 uzemi_cis  uzemi_cis  "kód číselníku pro referenční území " TRUE     string  
#> 10 uzemi_kod  uzemi_kod  "kód položky číselníku pro referenčn… TRUE     string  
#> 11 uzemi_txt  uzemi_txt  "text položky z číselníku pro refere… TRUE     string  
#> 12 stapro_txt stapro_txt "text statistické proměnné"           TRUE     string  
#> 13 spkvantil… spkvantil… "text položky číselníku pro kvantil"  TRUE     string  
#> 14 pohlavi_t… pohlavi_t… "text položky číselníku pro pohlaví"  TRUE     string  

# }